pyproject.toml
[project]
name = "gmailautomater"
version = "0.1.0"
description = ""
authors = [{ name = "Jayden Pyles", email = "jpylesbusiness@gmail.com" }]
dependencies = [
    "python-dotenv",
    "click>=8.1.7",
    "rich>=13.5.2",
    "imap-tools>=1.2.0",
    "pipx>=1.2.0",
]
requires-python = ">=3.11"
license = { text = "MIT" }

[tool.pdm]
[tool.pdm.scripts]
gmailautomater = "gmailautomater.__main__:main"

[tool.pyright]
include = ["src"]
exclude = ["**/node_modules", "**/__pycache__"]
ignore = []
defineConstant = { DEBUG = true }
stubPath = ""

reportMissingImports = true
reportMissingTypeStubs = false

pythonVersion = "3.9"
pythonPlatform = "Linux"


[tool.isort]
length_sort = "1"
profile = "black"
sections = "STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
import_heading_stdlib = "STL"
import_heading_thirdparty = "PDM"
import_heading_firstparty = "LOCAL"
import_heading_localfolder = "LOCAL"


[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
